requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6

# Visualization
matplotlib==3.8.2

# AI/ML
openai==1.3.7

# Document Generation
python-docx==1.1.0
reportlab==4.0.7
Pillow==10.1.0
PyPDF2==3.0.1

# HTTP Requests
requests==2.31.0

# Environment Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0